Replace old coordinator dashboard with new coordinator dashboard
